requirements.txt
langchain
langchain-community
langchain-openai
sentence-transformers
numpy
jsonschema
python-dotenv
